+++ Tests/Functional/FrontEnd/DefaultController/SingleViewTest.php
@@ -68,7 +68,7 @@ final class SingleViewTest extends FunctionalTestCase
 
         $frontEndController = $this->getFrontEndController();
         $subject = new TestingDefaultController();
-        $subject->cObj = $frontEndController->cObj;
+        $subject->setContentObjectRenderer($frontEndController->cObj);
         $subject->init(
             [
                 'isStaticTemplateLoaded' => 1,